+++ server/tests/peertube-runner/vod-transcoding.ts
@@ -334,8 +334,10 @@ describe('Test VOD transcoding in peertube-runner program', function () {
   })
 
   after(async function () {
-    await peertubeRunner.unregisterPeerTubeInstance({ server: servers[0] })
-    peertubeRunner.kill()
+    if (peertubeRunner) {
+      await peertubeRunner.unregisterPeerTubeInstance({ server: servers[0] })
+      peertubeRunner.kill()
+    }
 
     await cleanupTests(servers)
   })
